l_0419/requirements.txt
flask
volcenginesdkarkruntime
python-dotenv
requests==2.31.0
Pillow>=9.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6